refactor(onboarding): extract page-path/label helpers to break FindSpotlight<->ChecklistSection cycle
Move ONBOARDING_PAGE_PATHS, itemLabel, itemTooltip, prerequisiteReason,
PAGE_META, and PAGE_ORDER out of ChecklistSection.tsx into the new leaf
module onboarding-labels.ts so neither ChecklistSection nor FindSpotlight
imports from the other to get those values.

ChecklistSection re-exports the four previously-public symbols so no
external caller (FindSpotlight tests, ChecklistSection.reason-key.test) needs
updating. FindSpotlight now imports from onboarding-labels directly.

Prerequisite for wiring madge --circular into CI (DS-22).
